+++ nb_risk/__init__.py
@@ -11,15 +11,15 @@ class NbriskConfig(PluginConfig):
     version = __version__
     author = "Renato Almeida de Oliveira Zaroubin"
     author_email = "renato.almeida.oliveira@gmail.com"
-    min_version = "4.1.0"
-    max_version = "4.1.99"
+    min_version = "4.3.0"
+    max_version = "4.3.99"
     required_settings = []
     default_settings = {
         "supported_assets": [
             "dcim.device",
             "virtualization.virtualmachine",
             "tenancy.tenant",
-            "dcim.site",
+            "dcim.location",
         ],
         "additional_assets": [],
         "proxies": {},