+++ CHANGELOG.md
@@ -4,3 +4,13 @@
 
 * Add case sensitive name constrain to Vulnerability model
 * Change the import path of get_plugin_config() to extras.plugins.utils [13368](https://github.com/netbox-community/netbox/commit/f5a1f83f9fa9d98c945d21eb0f7ccb8cd37fbf59)
+
+## Version 43.0.0 (2023-05-14)
+* Added compatibility with NetBox 4.3.x
+* Updated models to include ordering
+* Replaced dcim.site with dcim.location for model references
+* Updated template extensions to use models instead of model
+* Fixed serializers to work with the latest NetBox API
+
+## Version 41.0.2
+* Added Netbox 4.1 Compatibility.